Fixes type error, makes compatible with playground
The model errors without the typing import.

Also, we need to provide a valid model name to use the model in the
playground. This provides a default model name that can be changed with
a runtime parameter.

When being called live, the model name can be adjusted. Or the default
can be used by calling the completion with model ==
"datarobot-deployed-llm".